Combining the size tests
